# stockone/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stock.one API",
        "description": "API de gestion de stocks pour papeteries - Benin",
        "contact": {
            "email": "admin@stockone.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://stockone.test:81/api/v1",
            "description": "Serveur local Laragon"
        }
    ],
    "paths": {
        "/alerts": {
            "get": {
                "tags": [
                    "Alertes"
                ],
                "summary": "Liste des alertes",
                "description": "Liste des alertes non lues",
                "operationId": "761f6279117c3162bb673a3db335196b",
                "parameters": [
                    {
                        "name": "type",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "unread_only",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alertes retournees"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/{id}/read": {
            "post": {
                "tags": [
                    "Alertes"
                ],
                "summary": "Marquer une alerte comme lue",
                "description": "Marquer une alerte comme lue",
                "operationId": "a4ffc73cfdf62ce75354aab77b68506a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/read-all": {
            "post": {
                "tags": [
                    "Alertes"
                ],
                "summary": "Marquer toutes les alertes comme lues",
                "description": "Marquer toutes les alertes comme lues",
                "operationId": "08a3d5f8599351c543066bbb781886a6",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/price-suggestions": {
            "get": {
                "tags": [
                    "Alertes"
                ],
                "summary": "Suggestions de reajustement de prix",
                "description": "Suggestions de prix IA",
                "operationId": "a46551e524bebcce4f203a8fd5b99d2a",
                "responses": {
                    "200": {
                        "description": "Suggestions retournees"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/price-suggestions/{id}/accept": {
            "post": {
                "tags": [
                    "Alertes"
                ],
                "summary": "Accepter une suggestion de prix",
                "description": "Accepter une suggestion de prix",
                "operationId": "bb331e082605e653a048cc9815bfc5ba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Prix mis a jour"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/alerts/price-suggestions/{id}/reject": {
            "post": {
                "tags": [
                    "Alertes"
                ],
                "summary": "Rejeter une suggestion de prix",
                "description": "Rejeter une suggestion de prix",
                "operationId": "47f63e4ce609fe8cc6d3682d6881ae41",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Suggestion rejetee"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}